split package hash
