+++ pam/fixtures/auth_01.conf
@@ -1,1 +1,1 @@
-wrapper_path = "./fixtures/response_01.sh"
+wrapper_path = "./fixtures/response_01"
