+++ pam/fixtures/auth_02.conf
@@ -1,1 +1,1 @@
-wrapper_path = "./fixtures/response_02.sh"
+wrapper_path = "./fixtures/response_02"
