+++ pam/fixtures/auth_03.conf
@@ -1,1 +1,1 @@
-wrapper_path = "./fixtures/response_03.sh"
+wrapper_path = "./fixtures/response_03"
