+++ pam/fixtures/auth_04.conf
@@ -1,1 +1,1 @@
-wrapper_path = "./fixtures/response_04.sh"
+wrapper_path = "./fixtures/response_04"
